LayoutTests/fast/canvas/webgl/tex-image-and-sub-image-2d-with-image-data-expected.txt

 
Verify texImage2D and texSubImage2D code paths taking ImageData, in particular flipY and premultiplyAlpha argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Regression test for https://bugs.webkit.org/show_bug.cgi?id=34459 : Refactor texImage2D and texSubImage2D taking ImageData to use common code
Testing texImage2D with flipY=true and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texImage2D with flipY=false and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texImage2D with flipY=true and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texImage2D with flipY=false and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=true and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=false and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=true and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=false and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
PASS successfullyParsed is true

TEST COMPLETE
